Linux fixes - using "__PRETTY_FUNCTION__" for "guardfunc" macro - using "THROW 1" instead of "THROW", otherwise appError didn't not work - added debug options to common.project (disabled with "!if 0" now) - libs/nvtt/nvcore/Memory.h now uses appMalloc etc for allocations - "operator new" in nvtt is FORCEINLINE now (NV_FORCEINLINE didn't work because it's just "__attribute__((always_inline))" without extra "inline"
Other:
- removed libs/nvtt/nvcore/Memory.cpp - not needed anymore
- added "operator delete[]" to Core.h

Earlier, before nvtt/Memory.cpp removal, there was a crash on umodel exit in memory allocator, caused long exception dump to console. When nvtt/Memory.cpp was removed and all memory functions were replaced with appMalloc etc, umodel crashed on startup in "operator delete" which was called by linux CRT code for own needs. That happened because nvtt's "operator delete" was not inlined (it was not marked as FORCEINLINE), so nvtt's operator delete became **global operator delete** - that was not good.