Fix _findnexti64() crashing on Win64 compiled by MSVC
I was getting crashes in the _findnexti64() call trying to access pointers in the style of 0x00000000XXXXXXXX (i.e. upper half of the word was zeroed). This made me suspicious that there's some pointer type truncation going on. Upon investigation, I've found that the return value of _findfirsti64() was being written to a long variable (which is always a 32-bit integer in MSVC), whereas the function returns an intptr_t (i.e. type equivalent in size to platform-specific pointer, and 64 bits in this case).