Fixed "abort during exception handling" error
Happend during appError handline, followed by program termination. There was a destructor in CTextContainer, which released buffer memory. This destructor called appFree() which attempted to lock GMallocMutex, which was already released (as a global variable).
Added guards showing if CMutex is alive or not, ignoring any Lock/Unlock if mutex is not ready.